news_backend/requirements.txt
asgiref==3.4.1
boto3==1.19.2
botocore==1.22.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.8
django-allauth==0.45.0
django-cors-headers==3.10.0
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.12.4
dnspython==2.1.0
gunicorn==20.1.0
idna==3.3
jmespath==0.10.0
Markdown==3.3.4
mongoengine==0.23.1
oauthlib==3.1.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.3.0
pymongo==3.12.1
python-dateutil==2.8.2
python-decouple==3.5
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
